Remove `buildMessage` API now accepts PB only
The "convenience" of accepting JSON not PB was not
worth it. It didn't really work and didn't allow nested
messages. Nor did it really conform to the PB<->JSON spec.

So we now only accept PB messages, nothing raw, even if the
transport layer uses JSON.

This came out of PR:
https://github.com/thechriswalker/protoc-gen-twirp_js/pull/2